dev_project/lib/backend/python_backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
supabase>=2.6,<3.0
python-dotenv>=1.0,<2.0
